pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]
name = "NOMADworkshop"
description = "test for a plugin"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Joshua Damm", email = "joshua.damm@kit.edu" },
]
maintainers = [
    { name = "Joshua Damm", email = "joshua.damm@kit.edu" },
]
license = { file = "LICENSE" }
dependencies = [
    "nomad-baseclasses @ git+https://github.com/nomad-hzb/nomad-baseclasses.git",
    "nomad-lab>=1.3.16",
    "python-magic-bin; sys_platform == 'win32'",
]

[project.urls]
Repository = "https://github.com/NOMADe-Joshua/NOMADworkshop"

[project.optional-dependencies]
dev = ["ruff", "pytest", "structlog"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4


[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # isort
    "I",
    # pylint
    "PL",
]

ignore = [
    "F403", # 'from module import *' used; unable to detect undefined names
    "PLR2004", # magic value used in comparison
    "E501", # line too long
    "PLR0915", # too-many-statements
    "PLR0912", # too-many-branches
]

fixable = ["ALL"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# this is entirely optional, you can remove this if you wish to
[tool.ruff.format]
# use single quotes for strings.
quote-style = "single"

# indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[project.entry-points.'nomad.plugin']
parser_entry_point = "nomadworkshop.parsers:parser_entry_point"
joshuashirn_experiment_parser_entry_point = "nomadworkshop.parsers:joshuashirn_experiment_parser_entry_point"
joshuashirn_parser_entry_point = "nomadworkshop.parsers:joshuashirn_parser_entry_point"
schema_package_entry_point = "nomadworkshop.schema_packages:schema_package_entry_point"
joshuashirn_schema_package_entry_point = "nomadworkshop.schema_packages:joshuashirn_schema_package_entry_point"
normalizer_entry_point = "nomadworkshop.normalizers:normalizer_entry_point"
app_entry_point = "nomadworkshop.apps:app_entry_point"
joshuashirn_voila_app = "nomadworkshop.apps:joshuashirn_voila_app"
example_upload_entry_point = "nomadworkshop.example_uploads:example_upload_entry_point"
voila_scripts_entry_point = "nomadworkshop.example_uploads:voila_scripts_entry_point"
[tool.cruft]
# Avoid updating workflow files, this leads to permissions issues
skip = [".github/*"]
